jUnitの勉強 jUnit実践入門 １章Tutorial 除算メソッド修正

+++ testFrameWork/jUnit/src/main/java/junit/tutorial/Calculator.java
@@ -5,7 +5,7 @@ public class Calculator {
 		return x * y;
 	}
 
-	public int divide(int x, int y) {
-		return x / y;
+	public float divide(int x, int y) {
+		return (float) x / (float) y;
 	}
 }
